flow-statistics: add average sent packet rate
The average sent packet rate can be determined from the received
packet numbers.  Since not all packets need to have the same size,
average bandwidth cannot be determined this way.

If the first or last packet is lost, the printed sent rate is wrong.

TCP, of course, messes this up.  All of mgen-tools has been developed
for use with UDP based traffic.